+++ x/launchpad/keeper/keeper.go
@@ -3,7 +3,6 @@ package keeper
 import (
 	"fmt"
 
-	gogotypes "github.com/gogo/protobuf/types"
 	"github.com/notional-labs/anone/v043_temp/address"
 	"github.com/tendermint/tendermint/libs/log"
 
@@ -54,86 +53,6 @@ func (k Keeper) GetModuleAccountAddress(ctx sdk.Context) sdk.AccAddress {
 }
 
 // ============ Project Helper Logic
-
-// GetNextProjectIDAndIncrement returns the next project id, and increments the corresponding state entry.
-func (k Keeper) GetNextProjectIDAndIncrement(ctx sdk.Context) uint64 {
-	var projectID uint64
-	store := ctx.KVStore(k.storeKey)
-
-	bz := store.Get(types.KeyNextGlobalProjectID)
-	if bz == nil {
-		panic(fmt.Errorf("project has not been initialized -- Should have been done in InitGenesis"))
-	} else {
-		val := gogotypes.UInt64Value{}
-
-		err := k.cdc.Unmarshal(bz, &val)
-		if err != nil {
-			panic(err)
-		}
-
-		projectID = val.GetValue()
-	}
-
-	k.SetNextProjectID(ctx, projectID+1)
-	return projectID + 1
-}
-
-func (k Keeper) GetProjectById(ctx sdk.Context, projectId uint64) (types.Project, error) {
-	store := ctx.KVStore(k.storeKey)
-	projectKey := types.GetKeyPrefixProject(projectId)
-	if !store.Has(projectKey) {
-		fmt.Errorf("project with ID %d does not exist", projectKey)
-		return types.Project{}, nil
-	}
-	project, err := k.UnmarshalProject(store.Get(projectKey))
-	if(err != nil) {
-		return types.Project{}, err
-	}
-
-	return project, nil
-}
-
-func (k Keeper) GetProjectAddress(ctx sdk.Context, projectId uint64) (sdk.AccAddress, error) {
-	store := ctx.KVStore(k.storeKey)
-	projectKey := types.GetKeyPrefixProject(projectId)
-	if !store.Has(projectKey) {
-		fmt.Errorf("project with ID %d does not exist", projectKey)
-		return sdk.AccAddress{}, nil
-	}
-	project, err := k.UnmarshalProject(store.Get(projectKey))
-	if(err != nil) {
-		return sdk.AccAddress{}, err
-	}
-	projectAddress := k.accountKeeper.GetModuleAddress(project.ProjectAddress)
-	return projectAddress, nil
-}
-
-func (k Keeper) GetAllProjects(ctx sdk.Context) (res []types.Project, err error) {
-	store := ctx.KVStore(k.storeKey)
-	iter := sdk.KVStorePrefixIterator(store, types.KeyPrefixProject)
-	defer iter.Close()
-	for ; iter.Valid(); iter.Next() {
-		bz := iter.Value()
-
-		project, err := k.UnmarshalProject(bz)
-		if err != nil {
-			return nil, err
-		}
-
-		//only get projects that have not been deleted
-		if(project != types.Project{}) {
-			res = append(res, project)
-		}
-	}
-	return res, nil
-}
-
-// SetNextProjectID sets next project ID.
-func (k Keeper) SetNextProjectID(ctx sdk.Context, projectID uint64) {
-	store := ctx.KVStore(k.storeKey)
-	bz := k.cdc.MustMarshal(&gogotypes.UInt64Value{Value: projectID})
-	store.Set(types.KeyNextGlobalProjectID, bz)
-}
 
 // Get new project address
 func (k Keeper) NewProjectAddress(projectID uint64) sdk.AccAddress {
