+++ x/launchpad/keeper/store.go
@@ -7,6 +7,7 @@ import (
 	"github.com/notional-labs/anone/x/launchpad/types"
 
 	"github.com/gogo/protobuf/proto"
+	gogotypes "github.com/gogo/protobuf/types"
 )
 
 func (k Keeper) MarshalProject(project types.Project) ([]byte, error) {
@@ -31,13 +32,82 @@ func (k Keeper) SetProject(ctx sdk.Context, project types.Project) error {
 	return nil
 }
 
-func (k Keeper) DeletePool(ctx sdk.Context, projectID uint64) error {
+// GetNextProjectIDAndIncrement returns the next project id, and increments the corresponding state entry.
+func (k Keeper) GetNextProjectIDAndIncrement(ctx sdk.Context) uint64 {
+	var projectID uint64
 	store := ctx.KVStore(k.storeKey)
-	projectKey := types.GetKeyPrefixProject(projectID)
-	if !store.Has(projectKey) {
-		return fmt.Errorf("pool with ID %d does not exist", projectKey)
+
+	bz := store.Get(types.KeyNextGlobalProjectID)
+	if bz == nil {
+		panic(fmt.Errorf("project has not been initialized -- Should have been done in InitGenesis"))
+	} else {
+		val := gogotypes.UInt64Value{}
+
+		err := k.cdc.Unmarshal(bz, &val)
+		if err != nil {
+			panic(err)
+		}
+
+		projectID = val.GetValue()
 	}
 
-	store.Delete(projectKey)
-	return nil
+	k.SetNextProjectID(ctx, projectID+1)
+	return projectID + 1
+}
+
+func (k Keeper) GetProjectById(ctx sdk.Context, projectId uint64) (types.Project, error) {
+	store := ctx.KVStore(k.storeKey)
+	projectKey := types.GetKeyPrefixProject(projectId)
+	if !store.Has(projectKey) {
+		fmt.Errorf("project with ID %d does not exist", projectKey)
+		return types.Project{}, nil
+	}
+	project, err := k.UnmarshalProject(store.Get(projectKey))
+	if(err != nil) {
+		return types.Project{}, err
+	}
+
+	return project, nil
+}
+
+func (k Keeper) GetProjectAddress(ctx sdk.Context, projectId uint64) (sdk.AccAddress, error) {
+	store := ctx.KVStore(k.storeKey)
+	projectKey := types.GetKeyPrefixProject(projectId)
+	if !store.Has(projectKey) {
+		fmt.Errorf("project with ID %d does not exist", projectKey)
+		return sdk.AccAddress{}, nil
+	}
+	project, err := k.UnmarshalProject(store.Get(projectKey))
+	if(err != nil) {
+		return sdk.AccAddress{}, err
+	}
+	projectAddress := k.accountKeeper.GetModuleAddress(project.ProjectAddress)
+	return projectAddress, nil
+}
+
+func (k Keeper) GetAllProjects(ctx sdk.Context) (res []uint64, err error) {
+	store := ctx.KVStore(k.storeKey)
+	iter := sdk.KVStorePrefixIterator(store, types.KeyPrefixProject)
+	defer iter.Close()
+	for ; iter.Valid(); iter.Next() {
+		bz := iter.Value()
+
+		project, err := k.UnmarshalProject(bz)
+		if err != nil {
+			return nil, err
+		}
+
+		//only get projects that have not been deleted
+		if(project != types.Project{}) {
+			res = append(res, project.ProjectId)
+		}
+	}
+	return res, nil
+}
+
+// SetNextProjectID sets next project ID.
+func (k Keeper) SetNextProjectID(ctx sdk.Context, projectID uint64) {
+	store := ctx.KVStore(k.storeKey)
+	bz := k.cdc.MustMarshal(&gogotypes.UInt64Value{Value: projectID})
+	store.Set(types.KeyNextGlobalProjectID, bz)
 }
